+++ engine/src/bmr/BufferProvider.h
@@ -28,11 +28,12 @@ class allocation_pool;
 
 
 struct blazing_allocation{
-
-    std::size_t size;
-    char *data;    
-    std::stack< std::unique_ptr<blazing_allocation_chunk> > allocation_chunks;
-    allocation_pool * pool;  // this is to know how to free
+    std::size_t index; // index inside the pool
+    std::size_t size; // the size in bytes of the allocation
+    std::size_t total_number_of_chunks; // number of chunks when originally created
+    char *data;    // the pointer to the allocated memory
+    std::stack< std::unique_ptr<blazing_allocation_chunk> > allocation_chunks; // These are the available chunks that are part of the allocation. 
+    allocation_pool * pool;  // this is the pool that was used to make this allocation, and therefore this is what we would use to free it
 };
 
 struct blazing_allocation_chunk{
@@ -140,9 +141,24 @@ private:
 std::pair< std::vector<ral::memory::blazing_chunked_column_info>, std::vector<std::unique_ptr<ral::memory::blazing_allocation_chunk> >> convert_gpu_buffers_to_chunks(
     std::vector<std::size_t> buffer_sizes,bool use_pinned);
 
-std::shared_ptr<allocation_pool > get_host_buffer_provider();
-std::shared_ptr<allocation_pool > get_pinned_buffer_provider();
+/**
+	@brief This class represents the buffer providers used by the engine.
+  @note Myers' singleton. Thread safe and unique. Note: C++11 required.
+*/
+class buffer_providers {
+public:
+  static std::shared_ptr<allocation_pool > & get_host_buffer_provider() {
+    static std::shared_ptr<allocation_pool> host_buffer_instance{};
+    return host_buffer_instance;
+  }
 
+  static std::shared_ptr<allocation_pool > & get_pinned_buffer_provider(){
+    static std::shared_ptr<allocation_pool> pinned_buffer_instance{};
+    return pinned_buffer_instance;
+  }
+};
+
+// this function is what originally initialized the pinned memory and host memory allocation pools
 void set_allocation_pools(std::size_t size_buffers_host, std::size_t num_buffers_host,
     std::size_t size_buffers_pinned, std::size_t num_buffers_pinned, bool map_ucx,
     ucp_context_h context);